requirements.txt
numpy~=1.19.5
pandas~=1.4.2
tensorflow-gpu==2.6.2
matplotlib~=3.5.1
plotly~=5.6.0
sklearn~=0.0
scikit-learn~=1.0.2
keras~=2.6.0
opencv-python~=4.5.5.64
Pillow~=9.1.0